Move project logic out of model
